% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f0_inicial.R
\name{f0_inicial}
\alias{f0_inicial}
\title{Función inicial}
\usage{
f0_inicial(directorio)
}
\arguments{
\item{directorio}{definir el directorio donde se crearan las carpetas.}
}
\description{
Funcion que instala las librerias necesarias para todo el proceso y crea las carpetas en donde se guardaran los archivos.
}
\details{
Se revisa si las funciones están instaladas en el entorno, en caso de
que no estén instaladas se procederán a instalar, luego se procede a revisar
si están creadas las carpetas donde se guardarán los archivos creados con la librería
en total se crearán siete carpetas, para cada una de las funciones que genera algún archivo de salida.

1 S1_integracion:\code{\link{f1_integracion}}

2 S2_estandarizacion:\code{\link{f2_estandarizacion}}

3 S3_identificacion_alertas:\code{\link{f3_identificacion_alertas}}

4 S4_imputacion:\code{\link{f4_imputacion}}

5 S5_tematica:\code{\link{f5_tematica}}

6 S6_anexos:\code{\link{f6_anacional}} y \code{\link{f7_aterritorial}}

7 S7_boletin:\code{\link{f8_boletin}}
}
\examples{
f0_inicial(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET")
}
