% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1_integracion.R
\name{f1_integracion}
\alias{f1_integracion}
\title{Integración}
\usage{
f1_integracion(directorio, mes, anio)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}
}
\value{
CSV file
}
\description{
Con esta función se integraran los archivos de entrada que son necesarios en el proceso EMMET, con el fin
 de trabajar con una base llamada "base panel.csv". Adicionalmente, en las variables numeticas, los datos faltantes se cambian por cero.
 A estas mismas variables se es establece un formato numerico para la porterior manipulación.
}
\details{
En esta función se leen las siguientes bases:

1 Base logística: Es la base con todos los datos desde enero de 2018 hasta el mes
 indicado en el parámetro.

2 Base paramétrica: Base con las características de cada establecimiento.


El procedimiento para unificar las bases es el siguiente:

1 Se cambia el formato de la base logistica de xlsx a csv renombrandola como base panel.

2 Se agregan las columnas de la base paramétrica a la base panel por medio del ID_NUMORD.
}
\examples{
f1_integracion(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022)
}
