% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2_alertas.R
\name{f2_identificacion_alertas}
\alias{f2_identificacion_alertas}
\title{Identificación de alertas}
\usage{
f2_identificacion_alertas(directorio, mes, anio, avance = 100)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada.}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}

\item{avance}{Denifir que porcentaje de la base va a ser procesada para detectar alertas, por defecto el valor
esta en 100, lo cual indica que estan los más de 3000 establecimientos para el procesamiento, este genera un
archivo llamado por ejemplo: "EMMET_PANEL_alertas_nov2022.csv", si el avance es diferente de 100 el nombre será
"EMMET_PANEL_alertas_nov2022_50.csv" indicando que es el 50\% de la base}
}
\value{
CSV file
}
\description{
Función para identificar que variables dentro de cada establecimiento  pueden llegar a ser casos
de imputación, ya sean imputación deuda o por casos especiales.
}
\details{
\itemize{

\item Imputación deuda: son todos los establecimientos cuya novedad en la base original tienen valor de 5
(deuda), luego después de la imputación el código de novedad pasa a ser 98 (imputado por deuda).

\item Casos especiales: son aquellos establecimientos de las cuales se detectó que el dato informado
no es lo esperado, se llama al establecimiento para confirmar que el dato reportado es el que es, en dado caso que no exista respuesta por parte de ellos se toma el dato como un caso especial y se procede a imputar.
}



Primero se dividen las variables en capítulo 2 y capítulo 3

\itemize{

\item Variables capítulo 2

-II_PA_PP__NPERS_EP        |   -AJU_II_PA_PP_SUELD_EP

-II_PA_TD__NPERS_ET        |   -AJU_II_PA_TD_SUELD_ET

-II_PA_TI__NPERS_ETA       |   -AJU_II_PA_TI_SUELD_ETA

-II_PA_AP__AAEP            |   -AJU_II_PA_AP_AAS_AP

-II_PP_PP__NPERS_OP        |   -AJU_II_PP_PP_SUELD_OP

-II_PP_TD__NPERS_OT        |   -AJU_II_PP_TD_SUELD_OT

-II_PP_TI__NPERS_OTA       |   -AJU_II_PP_TI_SUELD_OTA

-II_PP_AP__APEP            |   -AJU_II_PP_AP_AAS_PP

-AJU_II_HORAS_HORDI_T      |    -AJU_II_HORAS_HEXTR_T

\item Variables capítulo 3

-AJU_III_PE_PRODUCCION

-AJU_III_PE_VENTASIN

-AJU_III_PE_VENTASEX

-III_EX__VEXIS

}
El procedimiento para identificar posibles establecimientos a imputar es:

1.	Si la novedad es igual 5 sabemos que el individuo se va a imputar por deuda en todas sus variables o si el valor reportado en la variable de interés es cero y el mes inmediatamente anterior fue diferente de cero también se identifica como imputación deuda

2.	Si la novedad es diferente a 5 se observa la variación con respecto al mes anterior; si es menor o igual a 20\% no se imputa.

3.	Si la variación es mayor a 20% y su valor es diferente de cero, entonces, primero se observa
si el dato es igual a alguno de los datos reportados por el establecimiento en meses anteriores;
en caso de que esté presente, se decide no imputar, si el dato no estuvo anteriormente, entonces
se procede a identificar si la variable es referente a capítulo 2, si es así,
se realiza una carta de control con los últimos 24 meses, si esta identifica el valor como atípico
se imputa por caso especial; si la variable es referente a capítulo 3, entonces, se realiza una prueba
mezclando el comando \code{\link[tsoutliers]{locate.outliers.iloop}} y la carta de control de 24 meses, el comando locate.outliers.iloop identifica los valores que
pueden ser outliers en series de tiempo, si el valor fue identificado como un valor atípico se imputa
por caso especial, en caso de que no fuera identificado como un valor atípico no se imputa.

Los límites de la carta control son calculados de la siguiente manera:
\deqn{(\bar{x}−1.96∗𝑆,\bar{x}+1.96∗𝑆)}

Donde \eqn{\bar{x}} es el promedio de la serie de los últimos 24 meses de la variable y establecimiento de interés y \eqn{S} es la desviación estándar de la serie de los últimos 24 meses de la variable y establecimiento de interés.
}
\section{Guia para actualizar archivo de alertas}{


Luego de ejecutar esta función se creara un archivo de excel tipo csv, este contiene variables
de identificación de los establecimientos, los valores en cada una de las variables de interes
(capitulo 2 y capitulo 3), y si son posibles casos de imputación, las variables en las que
tengan un valor diferente a continua son los que en la función de imputación pasaran por ese proceso

Para modificar el archivo debe modifcar el valor en la variable que desee y busque la columna
cuyo nombre es nombrevariable_caso_de_imputacion y modifique el valor por "continua_critica",para
tener registro de que valores se modificaron por una critica, si desea, en la ultima columna
puede realizar un comentario

Ejemplo, queremos modificar en valor de la producción del establecimiento con ID_numord 23;
por lo tanto primero se busca la fila cuyo valor de id_numord es 23, en la columna "AJU_III_PE_PRODUCCION"
cambiaremos el valor numerico por el valor que deseamos (evite usar decimales), luego proceda
a buscar la columna "AJU_III_PE_PRODUCCION_caso_de_imputacion", ahi modifique el valor de la casilla
por "continua_critica", sin importar si el valor anterior era "continua", "imputacion_deuda"
 o "imputacion_caso_especial".
}

\examples{
f2_identificacion_alertas(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022,avance=100)

}
