% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3_imputacion.R
\name{f3_imputacion}
\alias{f3_imputacion}
\title{Imputación outliers}
\usage{
f3_imputacion(directorio, mes, anio, avance = 100)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir las tres primeras letras del mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}

\item{avance}{Denifir que porcentaje de la base va a ser cargada para el proceso de imputación, por defecto el valor
esta en 100.}
}
\value{
CSV file
}
\description{
Función para realizar la imputación por deuda y por casos especiales.
}
\details{
La metodología de imputación es la siguiente:

1.	Para las variables de capítulo 2 la metodología a usar es la imputación por el mes anterior.

2.	Para las variables de capítulo 3 la metodología a usar para imputación por casos especiales
 es el KNN combinado con variación mes anterior. Para la imputación por deuda en variables pertenecientes
 a capítulo 3 se realiza un proceso similar.

El KNN combinado consta de los siguientes pasos:

1. Realizar un primer KNN imputando los individuos atípicos en las variables de interés (KNN1).

2. Calcular la variación con respecto al mes anterior.

3. Realizar un segundo KNN para la variación con respecto al mes anterior de los
establecimientos del mismo dominio (KNN2).

4. Calcular el valor final de la siguiente manera:
\deqn{ \hat{y} = KNN_1*(1+KNN_2)}
Donde \eqn{ \hat{y}} es el valor resultante con el que se imputara la variable de interés.

Para el caso de imputación deuda en capítulo 3 se sigue la siguiente formula:
\deqn{ \hat{y} = MA*(1+KNN_2)}
Esto es que el valor resultante con el que se imputara la variable de interés es igual al valor
del mes anterior, multiplicado por 1 más la variación con respecto al mes anterior de los establecimientos
del mismo dominio.

Para conocer más sobre la imputación por el método k-Nearest Neighbors ver \code{\link[VIM:kNN]{KNN_VIM}}.
}
\examples{
f3_imputacion(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022,avance=100)
}
