% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f5_anexo_nacional.R
\name{f5_anacional}
\alias{f5_anacional}
\title{Anexo Nacional}
\usage{
f5_anacional(directorio, mes, anio)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}
}
\value{
CSV file
}
\description{
Esta funcion crea el archivo de Anexo Nacional.
 Tiene como insumo, la base de datos Tematica. El cuerpo de la función
 crea cada una de las hojas del reporte.

 Los cuadros de salida o anexos estadísticos de EMMET, muestran información
 complementaria a la registrada en el boletín de prensa con el fin de brindar
 la información a un nivel más desagregado tanto total nacional como
 desagregada a nivel de departamentos, áreas metropolitanas y principales
 ciudades del país.

 Los resultados se muestran con variaciones anuales, año corrido y doce meses,
 junto con sus respectivas contribuciones, según dominios, por las principales
 variables que se recolectan en el proceso: Producción (real y nominal),
 ventas (real y nominal) y empleo, desagregando a su vez esta última variable
 por área funcional y tipo de contrato.
 De igual manera se presentan los resultados de sueldos causados y horas
 totales trabajadas para los dominios nacionales.
}
\details{
En los anexos nacionales se calculan las contribuciones y variaciones
 en tres diferentes periodos.

 Es importante resaltar que el equipo en el que se ejecutará la función debe
 tener intalado Java. Ver \href{https://github.com/NataliArteaga/DANE.EMMET#readme}{README}

 A continuación se muestran los  periodos
 y las formulas para realizar el cálculo de estos:

 Contribución anual:

  \deqn{
 CA_{trj} = \frac{(V_{trj} - V_{(t-12)rj)}}{\sum_{1}^n V_{(t-12)rj)}} *100
 }


 Donde:

 t: Mes de referencia de la publicación de la operación estadística

 \eqn{V_{trj}}: Valor en el periodo t para el territorio r en el dominio j.

 \eqn{V_{(t-12)rj}}: Valor en el periodo t-12 o en el año anterior, en el
 territorio r en el dominio j.

 \eqn{\sum_{1}^n V_{(t-12)rj)}}: Sumatoria de los valores en el periodo t-12,
 en el territorio r y en el dominio j

 Esta contribución anual se interpreta como el aporte del domino j
 en el territorio r a la variación anual del mes de referencia en el
 domino j en el territorio r

 Contribución año corrido:

 \deqn{
 CAC_{Trj} = \frac{\sum_{i}^T(V_{trj} - \sum_{b}^{T-12} V_{trj)}}{\sum_{b}^{T-12} V_{trj)}} *100
 }


 Donde:
 t: Mes variando de enero a diciembre

 T: Mes de referencia.

 i: Siempre es el mes de enero.

 b=i-12: Corresponde a enero del año anterior

 \eqn{V_{trj}}: Valor de la variable en el periodo t en el territorio r
 en el dominio j

 Esta contribucion de año corrido se interpreta como el aporte del domino j
 en el territorio r a la variación año corrido del mes de referencia en el
 domino j en el territorio r.


 Contribución año acumulado:

 \deqn{
 CAA_{Trj} = \frac{\sum_{a+1}^T(V_{trj} - \sum_{b+1}^{a} V_{trj)}}{\sum_{b+1}^{a} V_{trj)}} *100
 }

 Donde:

 t: Mes variando de enero a diciembre

 T: Mes de referencia.

 a=T-12

 b=a-12: Corresponde al mes a del año anterior

 \eqn{V_{trj}}: Valor de la variable en el periodo t en el territorio r en el
 dominio j

 Nota: cuando las variables que denotan meses (a, b) son negativas representan
 el mes del año inmediatamente anterior.

 Esta contribución de año acumulado se interpreta como el aporte del domino
 j en el territorio r a la variación acumulada anual del mes de referencia
 en el domino j en el territorio r.


 Variación anual:

 Es la relación del índice o valor (para producción y ventas, categoría de
 contratación, sueldos, horas) en el mes de referencia (ti) con el índice o
 valor absoluto del mismo mes en el año anterior (t^{i-12}), menos 1 por 100.

 \deqn{
 VA = \frac{\text{índice o valor del mes de referencia}}
 {\text{índice o valor del mismo mes del año anterior}} -1 *100
 }

 Se interpreta como el crecimiento o disminución porcentual, dependiendo de
 si el resultado es negativo o positivo, de la variable correspondiente en
 el mes de referencia, en relación al mismo mes del año anterior


 Variación Año Corrido:

 \deqn{
 VAC = \frac{\sum \text{índice o valor de enero al mes de referencia del año actual}}
 {\sum \text{índice o valor de enero al mes de referencia del mismo mes del año anterior}} -1 *100
 }

 Se interpreta como el crecimiento o disminución porcentual, dependiendo de
 si el resultado es negativo o positivo, de la variable correspondiente en lo
 corrido del año hasta el mes de referencia, en relación al mismo periodo del
 año anterior


 Variación Acumulado Anual:

\deqn{
 VAA = \frac{\sum \text{índice o valor desde} a_{+1} \text{hasta el mes de referencia}}
 {\sum \text{índice o valor en el año anterior desde} a_{+1} \text{ hasta el mes de referencia}} -1 *100
 }

 Donde:

 t=mes de referencia

 a=t-12

 Se interpreta como el crecimiento o disminución porcentual, dependiendo de
 si el resultado es negativo o positivo, de la variable correspondiente en
 los últimos 12 meses hasta el mes de referencia, en relación al mismo periodo
 del año anterior.

 Contribuciones porcentuales: aporte en puntos porcentuales de las variaciones
 individuales a la variación de un agregado.


 La función escribe, en formato excel, las hojas:

 1. Var y cont_Anual:
 Se caclcula la Variación anual (%) y contribución, del valor
 de la producción, ventas, y empleo, según las clase industrial

 2. Var Anual_Emp_Sueldos_Horas:
 Se caclculan las Variaciones anuales, según clase industrial,
 producción, ventas, personal ocupado, sueldos y horas totales
 trabajadas

 3. Var y cont_año corrido:
 Se caclcula la Variación año corrido (%) y contribución, del valor de
 la producción, ventas, y empleo, según las clase industrial

 4. Var año corr Emp_Sueldos_Hor:
 Se caclculan las variaciones año corrido, según clase industrial
 producción, ventas, personal ocupado, sueldos y horas totales
 trabajadas


 5. Var y cont_doce meses:
 Se caclcula la  aariación doce meses (%) y contribución, del valor
 de la producción, ventas, y empleo, según las clase industrial


 6. Var doce meses Emp_Sueldos_H:
 Se caclculan las variaciones doce meses, según clase industrial,
 producción, ventas, personal ocupado, sueldos y horas totales trabajadas .


 7. Indices total por clase:
 Se calculan los índices de producción nominal y real, ventas nominal y real,
 empleo, sueldos y  horas Totales trabajadas, según clase industrial para
 cada uno de los dominios que se encuentran en la encuesta mensual
 manufacturera.

 8. Índices Desestacionalizados:
 Para la contrucción de la desestacionalización se usa la librería: seasonal.
 Esta libreria es una adaptacion del programa X13 seasonal, para el programa
 de R.

 Para esta función se crea un calendario en donde se ingresan los festivos
 fijos y los de fecha variables, para asignar la ponderación a los días, en
 donde finalmente quedan expresadas los días de Lunes a Sábados.

 Posteriormente para realizar la desestacionalización en las variables de
 producción, ventas y empleo, se escoge de manera individual la variable de
 interés con la columna de las fechas, este nuevo data frame
 (ej: producción, fecha) se convierte en una serie de tiempo, con frecuencia
 mensual.

 Para la desestacionalización, se usa la serie de tiempo construida
 anteriormente, y se especifican:

 * fecha de inicio y fin de la serie (ej: 2001.1,2021.11).

 * variables regresoras: el calendario, construido anteriormente,

 * variables dicotomicas de outliers: estas son festivos patrios,
      años bisiesto, y fechas de alto impacto, como:

      2016.Jul: Paro camionero

      2016.Aug:

      2020.Mar: Inicio de pandemia Covid-19

      2020.Apr: Inicio de la cuarentena

      2021.May: Paro Nacional

 Adicionalmente se especifican los parámetros estacionales y la informacion
 de la serie que se desea observar.

 Finalmente se crea una tabla con la fecha y los datos desestacionalizados.

 Esto, por cada una de las variables de interés.


 9. Enlace legal hasta 2014:
 Se realiza una conexión entre la serie de la metodología actual, con las
 variaciones de los indices de la metedología anterior con el fin de tener una
 serie más larga de comparación de la información.


 10. Enlace legal hasta 2001:
 Se realiza una conexión entre la serie de la metodología actual, con las
 variaciones de los indices de la metedología anterior con el fin de tener una
 serie más larga de comparación de la información.

 11. Var y cont_Trienal:
 Se caclcula la contribución y variación trienal, es decir usando como año
 base los datos del año 2019, del valor de la producción, ventas, y empleo,
 según clase industrial.
}
\examples{
f5_anacional(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022)


}
