% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7_cdominios.R
\name{f7_cdominios}
\alias{f7_cdominios}
\title{Cuadros desagregaion dominios}
\usage{
f7_cdominios(directorio, anio, mes)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{anio}{Definir el año a ejecutar, ej: 2022}

\item{mes}{Definir el mes a ejecutar, ej: 11}
}
\value{
CSV file
}
\description{
Esta funcion crea un archivo de excel que contiene las variaciones
y contribuciones anueales desagregadas por dominio 39. Por cada dominio se crea
una hoja en donde se consigna las variaciones y contribuciones de las variables
de produccion, ventas, emppleados, sueldos y horas.

Adicionalmente se crea una hoja en donde se presentan los valores de defalctores
y se calcula la variación anual y mensual por cada dominio.
}
\details{
Esta funcion se hace con la finalidad de crear cuadros informativos 
con datos desagregada según el interés, para facilitar el análisis del comportamiento
de cada secotr del dominio.
}
\examples{
f7_cdominios(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022)

}
