% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f9_boletin.R
\name{f9_boletin}
\alias{f9_boletin}
\title{Boletin}
\usage{
f9_boletin(directorio, mes, anio, tipo = "word")
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}

\item{tipo}{Definir si se quiere un archivo pdf, word, o html, por defecto esta la opción "word", ej: "html}
}
\description{
Está función crea el boletín, un documento exportado en pdf o word, según se desea.

 Este es un resumen ejecutivo en el que se presentan los
 principales resultados con el uso de cuadros de líneas para presentar
 información histórica y graficas de barras para hacer comparaciones de
 variaciones entre las distintas actividades o desagregación regional.
 Además, se presentan tablas con el resumen por actividad para el total
 nacional, departamentos, áreas metropolitanas y ciudades incluyendo
 las tres variables principales que permiten hacer un análisis cruzado
 entre estas.
}
\details{
Esta función requiere del uso de Google Chrome, pues para la creación de
las tablas de resumen en las que se muestra la contribución y
variación desagregada por total nacional, departamentos, áreas
metropolitanas y ciudades, se acuede a código html que compila gracias a
Google Chrome.
Las diferentes gráficas o tablas que se muestran, se realizan con las formulas
ya especificadas en ls detalles de las funciones de anexos.
Si requiere más información acerca de las fórmula, recurra a los detalles
de la función de anexos.
}
\examples{
f9_boletin(directorio"Documents/DANE/Procesos DIMPE /PilotoEMMET",mes = 11  ,anio =2022,tipo="pdf")

}
