% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro1.R
\name{macro1}
\alias{macro1}
\title{Macro 1}
\usage{
macro1(directorio, mes, anio, avance = 100)
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}

\item{avance}{Denifir que porcentaje de la base va a ser procesada para detectar alertas, por defecto el valor
esta en 100, lo cual indica que estan los más de 3000 establecimientos para el procesamiento, este genera un
archivo llamado por ejemplo: "EMMET_PANEL_alertas_nov2022.csv", si el avance es diferente de 100 el nombre será
"EMMET_PANEL_alertas_nov2022_50.csv" indicando que es el 50\% de la base}
}
\description{
Con esta función realizaremos la primera parte del proceso, en donde se correran las 3
primeras funciones.
}
\details{
Esta funcion tiene como objetivo correr desde la función inicial hasta la función
de identificacion de alertas, con el fin de que el usuario pueda obtener rapidamente el
archivo de salida de la función de alertas y poder realizar el proceso de critica.
Recuerde que si el archivo de alertas ya está creado, al correr esta función le pedira que
responda S si quiere sobreescribirlo, de lo contrario otro valor para cancelar la función

Ver:\code{\link{f0_inicial}},\code{\link{f1_integracion}},
\code{\link{f2_identificacion_alertas}}
}
\examples{
 macro1(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022,avance=100)
}
