% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro2.R
\name{macro2}
\alias{macro2}
\title{Macro 2}
\usage{
macro2(directorio, mes, anio, avance = 100, tipo = "pdf")
}
\arguments{
\item{directorio}{definir el directorio donde se encuentran ubicado los datos de entrada}

\item{mes}{Definir el mes a ejecutar, ej: 11}

\item{anio}{Definir el año a ejecutar, ej: 2022}

\item{avance}{Denifir que porcentaje de la base va a ser cargada para el proceso de imputación, por defecto el valor
esta en 100.}

\item{tipo}{Definir si se quiere un archivo pdf, word, o html, por defecto esta la opción "pdf", ej: "html}
}
\description{
Con esta función realizaremos la segunda parte del proceso, en donde se correran las 5
funciones restantes, para finalmente optener el boletin
}
\details{
Esta funcion tiene como objetivo correr desde la función de imputacion hasta la función
que crea el archivo de boletin, con el fin de que el usuario pueda continuar con el proceso
luego de haber realizado el proceso de critica al archivo de alertas, para imputar solo los
valores necesarios. Esta función le pedira una confirmación de que ya valido el archivo de
alertas, marque 1 para que la función corra, marque otro valor para cancelar.

Ver:\code{\link{f3_imputacion}}, \code{\link{f4_tematica}}, \code{\link{f5_anacional}},
\code{\link{f6_aterritorial}}, \code{\link{f7_boletin}}
}
\examples{
macro2(directorio="Documents/DANE/Procesos DIMPE /PilotoEMMET",
                       mes=11,anio=2022,avance=100,tipo="pdf")
}
